+++ apps/electron/src/main/application/agent/development/development-agent.ts
@@ -14,7 +14,7 @@ import {
   fetchServerResources,
   readServerResource,
   substituteArgsParameters,
-} from "@/main/domain/mcp-core/client/mcp-client-util";
+} from "@/main/application/mcp-core/client/mcp-client-util";
 
 /**
  * サーバー情報を表す型
