+++ apps/electron/src/main/application/agent/shared/agent-base.ts
