+++ apps/electron/src/main/application/agent/shared/agent-sharing-service.ts
