+++ apps/electron/src/main/application/agent/shared/agent-utils.ts
@@ -1,4 +1,4 @@
-import { extractServerVariables } from "@/main/domain/mcp-core/server/shared/server-variable-utils";
+import { extractServerVariables } from "@/main/application/mcp-core/server/shared/server-variable-utils";
 import { AgentConfig, DeployedAgent } from "@mcp_router/shared";
 
 /**
