+++ apps/electron/src/main/application/auth/auth.ts
