+++ apps/electron/src/main/application/mcp-core/client/mcp-app-paths.ts
