+++ apps/electron/src/main/application/mcp-core/client/mcp-client-util.ts
