+++ apps/electron/src/main/application/mcp-core/package/install-package-manager.ts
