+++ apps/electron/src/main/application/mcp-core/package/package-version-resolver.ts
