+++ apps/electron/src/main/application/mcp-core/server/server-service.ts
