+++ apps/electron/src/main/application/mcp-core/server/shared/server-variable-utils.ts
