+++ apps/electron/src/main/application/mcp-core/token/token-service.ts
