+++ apps/electron/src/main/application/workflow/hook-service.ts
