+++ apps/electron/src/main/application/workspace/workspace-service.ts
