+++ apps/electron/src/main/infrastructure/ipc/handlers/package-version-handlers.ts
@@ -9,7 +9,7 @@ import { ipcMain } from "electron";
 import {
   resolvePackageVersionsInArgs,
   checkMcpServerPackageUpdates,
-} from "@/main/domain/mcp-core/package/package-version-resolver";
+} from "@/main/application/mcp-core/package/package-version-resolver";
 
 /**
  * Register the IPC handlers for package version resolution
