+++ apps/electron/src/main/infrastructure/ipc/handlers/token-handler.ts
@@ -1,5 +1,5 @@
 import { ipcMain } from "electron";
-import { getTokenService } from "@/main/domain/mcp-core/token/token-service";
+import { getTokenService } from "@/main/application/mcp-core/token/token-service";
 import { TokenScope } from "@mcp_router/shared";
 
 export function setupTokenHandlers(): void {
